move prompts from create module to app-config-writer
