+++ packages/app-config-writer/src/index.ts
@@ -1,4 +1,4 @@
-export { getSmartLinksTargetFromPrompt } from './prompt';
+export { getSmartLinksTargetFromPrompt, simulatePrompt, includeTestRunnersPrompt } from './prompt';
 export { generateSmartLinksConfig } from './smartlinks-config';
 export { generateInboundNavigationConfig, readManifest } from './navigation-config';
 export { generateVariantsConfig } from './variants-config';
