+++ packages/app-config-writer/src/prompt/index.ts
@@ -1,1 +1,2 @@
 export { getSmartLinksTargetFromPrompt, promptUserPass } from './smartlinks-config';
+export * from './preview-config';
